+++ paddle_billing/Client.py
@@ -105,10 +105,6 @@ class Client:
 
     @staticmethod
     def serialize_json_payload(payload: dict) -> str:
-        # Removes unneeded level of nested CustomData data
-        if payload.get('custom_data') and 'data' in payload['custom_data']:
-            payload['custom_data'] = payload['custom_data']['data']
-
         json_payload = json_dumps(payload, cls=PayloadEncoder)
         final_json   = json_payload if json_payload != '[]' else '{}'