+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Check our main [developer changelog](https://developer.paddle.com/?utm_source=dx
 - Added `product` to `subscription.items[]`, see [related changelog](https://developer.paddle.com/changelog/2024/subscription-items-product?utm_source=dx&utm_medium=paddle-python-sdk)
 - Support custom prices when updating and previewing subscriptions, see [related changelog](https://developer.paddle.com/changelog/2024/add-custom-items-subscription)
 
+### Changed
+
+- `paddle_billing.Entities.Shared.CustomData` is no longer a `dataclass`
+
 ## 0.2.2 - 2024-09-03
 
 ### Fixed